+++ cpp_version/pmugg/shape/vec3.h
@@ -10,46 +10,46 @@ class Vec2;
 
 class Vec3 {
 public:
-    Vec3(float x = 0, float y = 0, float z = 0);
+    Vec3(double x = 0, double y = 0, double z = 0);
     ~Vec3() = default;
 
-    float getX() const { return x; }
-    float getY() const { return y; }
-    float getZ() const { return z; }
-    float getValue(int dim) const;
-    void setValue(float value, int dim);
-    float operator[](int dim) const { return getValue(dim); }
+    double getX() const { return x; }
+    double getY() const { return y; }
+    double getZ() const { return z; }
+    double getValue(int dim) const;
+    void setValue(double value, int dim);
+    double operator[](int dim) const { return getValue(dim); }
 
     Vec3& add(const Vec3& v);
     Vec3& minus(const Vec3& v);
-    Vec3& scale(float s);
+    Vec3& scale(double s);
     Vec3& normalize();
-    float dot(const Vec3& v) const;
+    double dot(const Vec3& v) const;
     Vec3 cross(const Vec3& v) const;
-    float length() const;
-    float length2() const;
-    float distance(const Vec3& v) const;
-    float distance2(const Vec3& v) const;
-    Vec3& rotate(const Vec3& axis, float theta);
-    Vec3& rotateX(float theta);
-    Vec3& rotateY(float theta);
-    Vec3& rotateZ(float theta);
+    double length() const;
+    double length2() const;
+    double distance(const Vec3& v) const;
+    double distance2(const Vec3& v) const;
+    Vec3& rotate(const Vec3& axis, double theta);
+    Vec3& rotateX(double theta);
+    Vec3& rotateY(double theta);
+    Vec3& rotateZ(double theta);
     Vec3 copy() const;
     Vec2 dropDim(int dim) const;
     std::string toString() const;
     static Vec3 import(const Json & json);
     Vec3 swapAxes() const;
 
-    static Vec3 lerp(const Vec3& start, const Vec3& end, float s);
-    static Vec3 unitVec(float theta, float phi);
-    static bool coordinatesClose(const Vec3& a, const Vec3& b, float tolerance);
+    static Vec3 lerp(const Vec3& start, const Vec3& end, double s);
+    static Vec3 unitVec(double theta, double phi);
+    static bool coordinatesClose(const Vec3& a, const Vec3& b, double tolerance);
 
     Vec3 operator+(const Vec3& v) const { return Vec3(x + v.x, y + v.y, z + v.z); }
     Vec3 operator-(const Vec3& v) const { return Vec3(x - v.x, y - v.y, z - v.z); }
-    Vec3 operator*(float s) const { return Vec3(x * s, y * s, z * s); }
+    Vec3 operator*(double s) const { return Vec3(x * s, y * s, z * s); }
     Vec3& operator+=(const Vec3& v) { return add(v); }
     Vec3& operator-=(const Vec3& v) { return minus(v); }
-    Vec3& operator*=(float s) { return scale(s); }
+    Vec3& operator*=(double s) { return scale(s); }
     bool operator==(const Vec3& v) const { return x == v.x && y == v.y && z == v.z; }
     bool operator!=(const Vec3& v) const { return !(*this == v); }
 
@@ -59,9 +59,9 @@ public:
     static const Vec3 Z_AXIS;
 
 private:
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 };
 
 } // namespace ms
